Fix misleading testing utilities and update tests
The deepEqual and isDeepEqual wrapper methods in test/util/assert.js
for some reason special-case string comparisons and use RegExp matching
in a way equivalent to String.prototype.includes.  This is confusing
and makes it very easy to write tests resulting in false positives.

This eliminates the alternate string behavior and updates the
tests that are broken by that change.

Fixes #94.