Force format=json and formatversion=2 in MW API requests
It's highly unlikely a user of this template would want a response in a
format other than JSON.  This hard-codes that into mwApiGet so that
consumers can skip the boilerplate and focus on the more relevant params
in their application code.

Updates the v1 siteinfo call to remove the boilerplate params added in
mwApiGet.

Also hard-codes formatversion=2 as we should be encouraging API consumers
to take advantage of this saner output format.